requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
jinja2
argon2-cffi